+++ src/constants.js
@@ -1,5 +1,5 @@
 // a string like `unauthenticated` will lead to the creation of a
-// constant named `EHA_USER_MANAGEMENT_AUTH_UNAUTHENTICATED_EVENT`
+// constant named `EHA_UMS_AUTH_UNAUTHENTICATED_EVENT`
 [
   'unauthenticated',
   'unauthorised'
@@ -7,8 +7,8 @@
   var upper = name.toUpperCase();
   // the value doesn't matter really, but having it like the name
   // might help troubleshooting
-  var constantNameAndValue = 'EHA_USER_MANAGEMENT_AUTH_'+upper+'_EVENT'
+  var constantNameAndValue = 'EHA_UMS_AUTH_'+upper+'_EVENT'
   angular
-    .module('eha.user-management-auth')
+    .module('eha.ums-auth')
     .constant(constantNameAndValue, constantNameAndValue);
 })