+++ src/http-interceptor.js
@@ -1,9 +1,9 @@
 ;(function() {
   'use strict';
 
-  var ngModule = angular.module('eha.user-management-auth.http-interceptor', []);
+  var ngModule = angular.module('eha.ums-auth.http-interceptor', []);
 
-  function EhaUserManagementAuthHttpInterceptor(options, $injector) {
+  function EhaUMSAuthHttpInterceptor(options, $injector) {
 
     function hostMatch(url) {
       var hosts = options.hosts;
@@ -22,28 +22,28 @@
 
     var $q = $injector.get('$q');
     var $log = $injector.get('$log');
-    var EHA_USER_MANAGEMENT_AUTH_UNAUTHENTICATED_EVENT = $injector.get('EHA_USER_MANAGEMENT_AUTH_UNAUTHENTICATED_EVENT');
+    var EHA_UMS_AUTH_UNAUTHENTICATED_EVENT = $injector.get('EHA_UMS_AUTH_UNAUTHENTICATED_EVENT');
 
     return {
       responseError: function(rejection) {
         // Check for 401 and hostMatch
         if (rejection.status === 401 && hostMatch(rejection.config.url)) {
-          var auth = $injector.get('ehaUserManagementAuthService');
-          auth.trigger(EHA_USER_MANAGEMENT_AUTH_UNAUTHENTICATED_EVENT);
+          var auth = $injector.get('ehaUMSAuthService');
+          auth.trigger(EHA_UMS_AUTH_UNAUTHENTICATED_EVENT);
         }
         return $q.reject(rejection);
       }
     };
   }
 
-  ngModule.provider('ehaUserManagementAuthHttpInterceptor', function() {
+  ngModule.provider('ehaUMSAuthHttpInterceptor', function() {
     var options = {};
     this.config = function(config) {
       options = config;
     };
 
     this.$get = function($injector) {
-      return new EhaUserManagementAuthHttpInterceptor(options, $injector);
+      return new EhaUMSAuthHttpInterceptor(options, $injector);
     };
   });